requirements.txt
beautifulsoup4
requests
pyyaml
lxml